Исправлена проблема при синхронизации миниатюр с последующим удалением Добавлена возможность синхронизации файлов по маске Добавлена возможность просмотра списка файлов на ручную синхронизацию
