+++ readme.txt
@@ -3,8 +3,8 @@ Contributors: Mauhem
 Donate link: http://wm-talk.net/supload-wordpress-plagin-dlya-zagruzki-na-selectel
 Tags: Uploads, Selectel, CDN, Cloud, Storage, media library, OpenStack, Object storage
 Requires at least: 3.0.1
-Tested up to: 4.0
-Stable tag: 1.2.3
+Tested up to: 4.1.1
+Stable tag: 1.3.0
 License: GPLv2
 License URI: http://www.gnu.org/licenses/gpl-2.0.html
